Stats save to phone
sky game is now mad ein MainActivity and asynchronously passed down to sky thread, this allows a listener between GameFragment and skyGame. GameFragment then saves the changes in stats using the saveStats function previously made by truman.

It should be fairly easy to add more stats to the game to be tracked. Just add a way to track it in skyGame or surfer, then add it to the save and read function in GameFragment and StatsFragment. Finally, update fragment_stats.xml with a new display.